+++ mf2.h
@@ -28,6 +28,8 @@ mf2_str_globals str_globals_mf2;
 
 #define MF2_STR(str) str_globals_mf2.str
 
+#define MF2_REGEX_ROOTS "/(?<=\\s|^)h-((?:[a-z0-9]+-)?[a-z]+(?:-[a-z]+)*)(?=\\s|$)/"
+
 zend_bool mf2_in_array( zval *haystack, zval *needle );
 int mf2_strcasecmp( const void *ida, const void *idb );
 void mf2_trim_html_space_chars( zval *trimmed_string, char *string );